refactor(web): extract ChannelRow component
